requirements.txt
python-dotenv==1.0.0
replicate==0.11.0
streamlit==1.22.0
streamlit-webrtc==0.45.0
